GEOMESA-2862 Split Avro/converter processors
* PutGeoMesaFoo for converters
* AvroToPutGeoMesaFoo for GeoAvro data files
