+++ geomesa-accumulo-bundle/geomesa-accumulo-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,1 +1,2 @@
 org.geomesa.nifi.processors.accumulo.PutGeoMesaAccumulo
+org.geomesa.nifi.processors.accumulo.AvroToPutGeoMesaAccumulo
