+++ geomesa-datastore-bundle/geomesa-datastore-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,2 +1,3 @@
-org.geomesa.nifi.datastore.processor.PutGeoTools
 org.geomesa.nifi.datastore.processor.ConvertToGeoAvro
+org.geomesa.nifi.datastore.processor.geotools.PutGeoTools
+org.geomesa.nifi.datastore.processor.geotools.AvroToPutGeoTools
