+++ geomesa-fs-bundle/geomesa-fs-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,1 +1,2 @@
 org.geomesa.nifi.processors.fs.PutGeoMesaFileSystem
+org.geomesa.nifi.processors.fs.AvroToPutGeoMesaFileSystem
