+++ geomesa-hbase-bundle/geomesa-hbase-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,1 +1,2 @@
 org.geomesa.nifi.processors.hbase.PutGeoMesaHBase
+org.geomesa.nifi.processors.hbase.AvroToPutGeoMesaHBase
