+++ geomesa-kafka-bundle/geomesa-kafka-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,1 +1,2 @@
 org.geomesa.nifi.processors.kafka.PutGeoMesaKafka
+org.geomesa.nifi.processors.kafka.AvroToPutGeoMesaKafka
