+++ geomesa-kafka-bundle/geomesa-kafka-processors/src/main/scala/org/geomesa/nifi/processors/kafka/PutGeoMesaKafka.scala
@@ -8,29 +8,6 @@
 
 package org.geomesa.nifi.processors.kafka
 
-import org.apache.nifi.annotation.behavior.InputRequirement.Requirement
-import org.apache.nifi.annotation.behavior.{InputRequirement, SupportsBatching}
-import org.apache.nifi.annotation.documentation.{CapabilityDescription, Tags}
-import org.apache.nifi.processor.ProcessContext
-import org.geomesa.nifi.datastore.processor.AbstractGeoIngestProcessor
-import org.geomesa.nifi.datastore.processor.utils.PropertyDescriptorUtils
-import org.locationtech.geomesa.kafka.data.KafkaDataStoreFactory.KafkaDataStoreFactoryParams
+import org.geomesa.nifi.datastore.processor.ConverterIngestProcessor
 
-@Tags(Array("geomesa", "kafka", "streaming", "stream", "geo", "ingest", "convert", "geotools"))
-@CapabilityDescription("Convert and ingest data files into GeoMesa")
-@InputRequirement(Requirement.INPUT_REQUIRED)
-@SupportsBatching
-class PutGeoMesaKafka extends AbstractGeoIngestProcessor(PutGeoMesaKafka.KafkaProperties) {
-  // set consumer count to zero to disable consuming
-  override protected def getDataStoreParams(context: ProcessContext): Map[String, _] =
-    super.getDataStoreParams(context) ++ Map(KafkaDataStoreFactoryParams.ConsumerCount.getName -> Int.box(0))
-}
-
-object PutGeoMesaKafka extends PropertyDescriptorUtils {
-
-  import KafkaDataStoreFactoryParams._
-
-  // note: KafkaDataStoreFactory.ParameterInfo is consumer-oriented, but we want producer properties here
-  private val KafkaProperties =
-    Seq(Brokers, Zookeepers, ZkPath, ProducerConfig, TopicPartitions, TopicReplication).map(createPropertyDescriptor)
-}
+class PutGeoMesaKafka extends GeoMesaKafkaProcessor with ConverterIngestProcessor
