+++ geomesa-redis-bundle/geomesa-redis-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,1 +1,2 @@
 org.geomesa.nifi.processors.redis.PutGeoMesaRedis
+org.geomesa.nifi.processors.redis.AvroToPutGeoMesaRedis
