+++ geomesa-redis-bundle/geomesa-redis-processors/src/main/scala/org/geomesa/nifi/processors/redis/AvroToPutGeoMesaRedis.scala
@@ -6,8 +6,8 @@
  * http://www.opensource.org/licenses/apache2.0.php.
  ***********************************************************************/
 
-package org.geomesa.nifi.processors.fs
+package org.geomesa.nifi.processors.redis
 
-import org.geomesa.nifi.datastore.processor.ConverterIngestProcessor
+import org.geomesa.nifi.datastore.processor.AvroIngestProcessor
 
-class PutGeoMesaFileSystem extends GeoMesaFileSystemProcessor with ConverterIngestProcessor
+class AvroToPutGeoMesaRedis extends GeoMesaRedisProcessor with AvroIngestProcessor
